+++ graphbrain/funs.py
@@ -1,5 +1,5 @@
-from graphbrain.memory.leveldb import LevelDB
+from graphbrain.memory.lmdb import LMDB
 
 
 def hypergraph(locator_string):
-    return LevelDB(locator_string)
+    return LMDB(locator_string)