+++ graphbrain/memory/lmdb.pyx
@@ -1,21 +1,22 @@
-import plyvel
+import shutil
+import lmdb
 from graphbrain.hypergraph import Hypergraph
 from graphbrain.hyperedge import *
 from graphbrain.memory.permutations import *
 
 
 def _ent2key(entity):
-    return (''.join(('v', entity.to_str()))).encode('utf-8')
+    return (''.join(('v', entity.to_str()))).encode()
 
 
 def _encode_attributes(attributes):
     str_list = [''.join((key, '|', str(attributes[key])))
                 for key in attributes]
-    return '\\'.join(str_list).encode('utf-8')
+    return '\\'.join(str_list)
 
 
 def _decode_attributes(value):
-    tokens = value.decode('utf-8').split('\\')
+    tokens = value.decode().split('\\')
     attributes = {}
     for token in tokens:
         parts = token.split('|')
@@ -23,47 +24,61 @@ def _decode_attributes(value):
     return attributes
 
 
-class LevelDB(Hypergraph):
-    """Implements LevelDB hypergraph storage."""
+class LMDB(Hypergraph):
+    """Implements LMDB hypergraph storage."""
 
     def __init__(self, locator_string):
         self.locator_string = locator_string
-        self.db = plyvel.DB(self.locator_string, create_if_missing=True)
+        self.env = lmdb.open(self.locator_string)
 
     # ============================================
     # Implementation of abstract interface methods
     # ============================================
 
     def close(self):
-        self.db.close()
+        pass
 
     def name(self):
         return self.locator_string
 
     def destroy(self):
-        self.db.close()
-        plyvel.destroy_db(self.locator_string)
-        self.db = plyvel.DB(self.locator_string, create_if_missing=True)
+        shutil.rmtree(self.locator_string)
+        self.env = lmdb.open(self.locator_string)
 
     def all(self):
-        start_str = 'v'
-        end_str = str_plus_1(start_str)
-        start_key = start_str.encode('utf-8')
-        end_key = end_str.encode('utf-8')
+        start_key = 'v'
+        end_key = str_plus_1(start_key)
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            yield hedge(key.decode('utf-8')[1:])
+        with self.env.begin() as txn:
+            cursor = txn.cursor()
+            if cursor.set_range(start_key.encode()):
+                search = True
+                key = cursor.key().decode()
+                while search and key < end_key:
+                    edge = hedge(key[1:])
+                    if cursor.next():
+                        key = cursor.key().decode()
+                    else:
+                        search = False
+                    yield edge
 
     def all_attributes(self):
-        start_str = 'v'
-        end_str = str_plus_1(start_str)
-        start_key = start_str.encode('utf-8')
-        end_key = end_str.encode('utf-8')
+        start_key = 'v'
+        end_key = str_plus_1(start_key)
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            edge = hedge(key.decode('utf-8')[1:])
-            attributes = _decode_attributes(value)
-            yield (edge, attributes)
+        with self.env.begin() as txn:
+            cursor = txn.cursor()
+            if cursor.set_range(start_key.encode()):
+                search = True
+                key = cursor.key().decode()
+                while search and key < end_key:
+                    edge = hedge(key[1:])
+                    attributes = _decode_attributes(cursor.value())
+                    if cursor.next():
+                        key = cursor.key().decode()
+                    else:
+                        search = False
+                    yield (edge, attributes)
 
     def atom_count(self):
         return self._read_counter('atom_count')
@@ -147,71 +162,108 @@ class LevelDB(Hypergraph):
                 positions.append(i)
         start_str = edges2str(nodes)
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('p', start_str))).encode('utf-8')
-        end_key = (''.join(('p', end_str))).encode('utf-8')
+        start_key = ''.join(('p', start_str))
+        end_key = ''.join(('p', end_str))
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            perm_str = key.decode('utf-8')
-
-            tokens = split_edge_str(perm_str[1:])
-            nper = int(tokens[-1])
-
-            if nper == first_permutation(len(tokens) - 1, positions):
-                edge = perm2edge(perm_str)
-                if edge and edge_matches_pattern(edge, pattern):
-                    yield edge
+        with self.env.begin() as txn:
+            cursor = txn.cursor()
+            if cursor.set_range(start_key.encode()):
+                search = True
+                key = cursor.key().decode()
+                while search and key < end_key:
+                    perm_str = key
+                    tokens = split_edge_str(perm_str[1:])
+                    if cursor.next():
+                        key = cursor.key().decode()
+                    else:
+                        search = False
+                    nper = int(tokens[-1])
+                    if nper == first_permutation(len(tokens) - 1, positions):
+                        edge = perm2edge(perm_str)
+                        if edge and edge_matches_pattern(edge, pattern):
+                            yield edge
 
     def _star(self, center, limit=None):
         center_str = center.to_str()
         start_str = ''.join((center_str, ' '))
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('p', start_str))).encode('utf-8')
-        end_key = (''.join(('p', end_str))).encode('utf-8')
+        start_key = ''.join(('p', start_str))
+        end_key = ''.join(('p', end_str))
 
-        count = 0
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            if limit and count >= limit:
-                break
-            perm_str = key.decode('utf-8')
-            edge = perm2edge(perm_str)
-            if edge:
-                position = edge.index(center)
-                nper = int(split_edge_str(perm_str[1:])[-1])
-                if nper == first_permutation(len(edge), (position,)):
-                    count += 1
-                    yield(edge)
+        with self.env.begin() as txn:
+            cursor = txn.cursor()
+            if cursor.set_range(start_key.encode()):
+                search = True
+                key = cursor.key().decode()
+                count = 0
+                while (search and
+                       key < end_key and
+                       (not limit or count < limit)):
+                    perm_str = key
+                    edge = perm2edge(perm_str)
+                    if cursor.next():
+                        key = cursor.key().decode()
+                    else:
+                        search = False
+                    if edge:
+                        position = edge.index(center)
+                        nper = int(split_edge_str(perm_str[1:])[-1])
+                        if nper == first_permutation(len(edge), (position,)):
+                            count += 1
+                            yield(edge)
 
     def _atoms_with_root(self, root):
         start_str = ''.join((root, '/'))
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('v', start_str))).encode('utf-8')
-        end_key = (''.join(('v', end_str))).encode('utf-8')
+        start_key = ''.join(('v', start_str))
+        end_key = ''.join(('v', end_str))
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            symb = hedge(key.decode('utf-8')[1:])
-            yield(symb)
+        with self.env.begin() as txn:
+            cursor = txn.cursor()
+            if cursor.set_range(start_key.encode()):
+                search = True
+                key = cursor.key().decode()
+                while search and key < end_key:
+                    symb = hedge(key[1:])
+                    if cursor.next():
+                        key = cursor.key().decode()
+                    else:
+                        search = False
+                    yield symb
 
     def _edges_with_edges(self, edges, root):
         start_str = ' '.join([edge.to_str() for edge in edges])
         if root:
             start_str = ''.join((start_str, ' ', root, '/'))
         end_str = str_plus_1(start_str)
-        start_key = (''.join(('p', start_str))).encode('utf-8')
-        end_key = (''.join(('p', end_str))).encode('utf-8')
+        start_key = ''.join(('p', start_str))
+        end_key = ''.join(('p', end_str))
 
-        for key, value in self.db.iterator(start=start_key, stop=end_key):
-            perm_str = key.decode('utf-8')
-            edge = perm2edge(perm_str)
-            if edge:
-                if root is None:
-                    if all([item in edge for item in edges]):
-                        positions = [edge.index(item) for item in edges]
-                        nper = int(split_edge_str(perm_str[1:])[-1])
-                        if nper == first_permutation(len(edge), positions):
+        with self.env.begin() as txn:
+            cursor = txn.cursor()
+            if cursor.set_range(start_key.encode()):
+                key = cursor.key().decode()
+                search = True
+                while search and key < end_key:
+                    perm_str = key
+                    edge = perm2edge(perm_str)
+                    if cursor.next():
+                        key = cursor.key().decode()
+                    else:
+                        search = False
+                    if edge:
+                        if root is None:
+                            if all([item in edge for item in edges]):
+                                positions = [edge.index(item)
+                                             for item in edges]
+                                nper = int(split_edge_str(perm_str[1:])[-1])
+                                if nper == first_permutation(len(edge),
+                                                             positions):
+                                    yield(edge)
+                        else:
+                            # TODO: remove redundant results when a root is
+                            # present
                             yield(edge)
-                else:
-                    # TODO: remove redundant results when a root is present
-                    yield(edge)
 
     def _set_attribute(self, edge, attribute, value):
         ent_key = _ent2key(edge)
@@ -250,12 +302,14 @@ class LevelDB(Hypergraph):
     def _add_key(self, ent_key, attributes):
         """Adds the given entity, given its key."""
         value = _encode_attributes(attributes)
-        self.db.put(ent_key, value)
+        with self.env.begin(write=True) as txn:
+            txn.put(ent_key, value.encode())
 
     def _write_edge_permutation(self, perm):
         """Writes a given permutation."""
-        perm_key = (''.join(('p', perm))).encode('utf-8')
-        self.db.put(perm_key, b'x')
+        perm_key = ''.join(('p', perm)).encode()
+        with self.env.begin(write=True) as txn:
+            txn.put(perm_key, b'x')
 
     def _write_edge_permutations(self, edge):
         """Writes all permutations of the edge."""
@@ -263,20 +317,23 @@ class LevelDB(Hypergraph):
 
     def _remove_edge_permutation(self, perm):
         """Removes a given permutation."""
-        perm_key = (''.join(('p', perm))).encode('utf-8')
-        self.db.delete(perm_key)
+        perm_key = ''.join(('p', perm))
+        with self.env.begin(write=True) as txn:
+            txn.delete(perm_key.encode())
 
     def _remove_edge_permutations(self, edge):
         """Removes all permutations of the edge."""
         do_with_edge_permutations(edge, self._remove_edge_permutation)
 
-    def _remove_key(self, ent_key):
-        """Removes an entity, given its key."""
-        self.db.delete(ent_key)
+    def _remove_key(self, edge_key):
+        """Removes an edge, given its key."""
+        with self.env.begin(write=True) as txn:
+            txn.delete(edge_key)
 
-    def _exists_key(self, ent_key):
+    def _exists_key(self, edge_key):
         """Checks if the given entity exists."""
-        return self.db.get(ent_key) is not None
+        with self.env.begin() as txn:
+            return txn.get(edge_key) is not None
 
     def _set_attribute_key(self, ent_key, attribute, value):
         """Sets the value of an attribute by ent_key."""
@@ -313,8 +370,9 @@ class LevelDB(Hypergraph):
             return False
 
     def _attribute_key(self, ent_key):
-        value = self.db.get(ent_key)
-        return _decode_attributes(value)
+        with self.env.begin() as txn:
+            value = txn.get(ent_key)
+            return _decode_attributes(value)
 
     def _get_str_attribute_key(self, ent_key, attribute, or_else=None):
         if self._exists_key(ent_key):
@@ -348,27 +406,30 @@ class LevelDB(Hypergraph):
 
     def __read_counter_key(self, counter_key):
         """Reads a counter by key."""
-        value = self.db.get(counter_key)
-        if value is None:
-            return 0
-        else:
-            return int(value.decode('utf-8'))
+        with self.env.begin() as txn:
+            value = txn.get(counter_key)
+            if value is None:
+                return 0
+            else:
+                return int(value)
 
     def _read_counter(self, counter):
         """Reads a counter by name."""
-        return self.__read_counter_key(counter.encode('utf-8'))
+        return self.__read_counter_key(counter.encode())
 
     def _inc_counter(self, counter, by=1):
         """Increments a counter."""
-        counter_key = counter.encode('utf-8')
-        value = self.__read_counter_key(counter_key)
-        self.db.put(counter_key, str(value + by).encode('utf-8'))
+        key = counter.encode()
+        value = self.__read_counter_key(key)
+        with self.env.begin(write=True) as txn:
+            txn.put(key, str(value + by).encode())
 
     def _dec_counter(self, counter, by=1):
         """Decrements a counter."""
-        counter_key = counter.encode('utf-8')
-        value = self.__read_counter_key(counter_key)
-        self.db.put(counter_key, str(value - by).encode('utf-8'))
+        key = counter.encode()
+        value = self.__read_counter_key(key)
+        with self.env.begin(write=True) as txn:
+            txn.put(key, str(value - by).encode())
 
     def _inc_degrees(self, edge, depth=0):
         if depth > 0: