+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.3.0] - ongoing work
+
+### Added
+- LMDB hypergraph backend.
+
+### Changed
+
+### Removed
+- LevelDB hypergraph backend.
+
 ## [0.2.2] - 13-09-2019
 
 ### Added